fix: move generateUrlFromEndpoint method to the client.go file
